Implement features

+++ app/Http/Controllers/Api/ProjectController.php
@@ -16,6 +16,7 @@ class ProjectController extends Controller
     public function create(Request $request)
     {
         try {
+            if (strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER') {
             //Validated
             $validateProject = Validator::make($request->all(), 
             [
@@ -32,12 +33,19 @@ class ProjectController extends Controller
 
             $project = Project::create([
                 'name' => $request->name,
+                'user_id' => auth()->user()->id,
             ]);
 
             return response()->json([
                 'status' => true,
                 'message' => 'Project Created Successfully',
             ], 200);
+        }
+        else{
+            return response([
+                'message' => 'Only PRODUCT_OWNER Role User Can Create Project.',
+            ]);
+        }
 
         } catch (\Throwable $th) {
             return response()->json([
@@ -49,8 +57,15 @@ class ProjectController extends Controller
     public function get(Request $request)
     {
         try {
-            $data = Project::find($request->id);
-
+           $data = [];
+            if(strtoupper(auth()->user()->userrole->role_name) == 'ADMIN') {
+            $data = Project::with('projectUser')->find($request->id);
+            }else if(strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER') {
+                $data = Project::with('projectUser')->where('user_id','=', auth()->user()->id)->find($request->id);
+             }else{
+               $data = [];
+             }
+               // print_r($data);
             if($data){
                 http_response_code(200);
                 return response([
@@ -76,7 +91,18 @@ class ProjectController extends Controller
     public function getAll(Request $request)
     {
         try {
-            $data = Project::orderby('id', 'desc')->get();
+            $data = [];
+            if(strtoupper(auth()->user()->userrole->role_name) == 'ADMIN') {
+            $data = Project::with('projectUser')->orderby('id', 'desc')->get();
+            }
+            else if(strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER') {
+                $data = Project::with('projectUser')->where('user_id','=', auth()->user()->id)->orderby('id', 'desc')->get();
+             }else{
+                return response([
+                    'message' => 'Unauthorized User!!',
+                ]);
+             }
+
 
             http_response_code(200);
             return response([
@@ -94,66 +120,79 @@ class ProjectController extends Controller
     }
     public function update(Request $request, $id)
     {
-        $validateProject = Validator::make($request->all(),
-            [
-                'name' => 'required',
-            ]);
-
-            if($validateProject->fails()){
-                return response()->json([
-                    'status' => false,
-                    'message' => 'validation error',
-                    'errors' => $validateProject->errors()
-                ], 401);
-            }
         try {
             $data = Project::findOrFail($id);
-            $data->name = $request->name;
-            $data->save();
+            if(strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER' && auth()->user()->id == $data->user_id) {
+                $validateProject = Validator::make($request->all(),
+                [
+                    'name' => 'required',
+                ]);
 
-            http_response_code(200);
-            return response([
-                'message' => 'Update Successful',
-            ]);
+                if($validateProject->fails()){
+                    return response()->json([
+                        'status' => false,
+                        'message' => 'validation error',
+                        'errors' => $validateProject->errors()
+                    ], 401);
+                }
+
+                    $data->name = $request->name;
+                    $data->save();
+
+                    return response([
+                        'message' => 'Update Successful',
+                    ]);
+             }else{
+                return response([
+                    'message' => 'Only Owner Can Update Project',
+                ]);
+             }
 
         } catch (RequestException $r) {
 
             http_response_code(400);
             return response([
-                'message' => 'Data failed to be updated.',
+                'message' => 'Data failed to be updated Project.',
                 'errorCode' => 4101,
             ], 400);
         }
     }
     public function patchupdate(Request $request, $id)
     {
-        $validateProject = Validator::make($request->all(),
-            [
-                'name' => 'required',
-            ]);
-
-            if($validateProject->fails()){
-                return response()->json([
-                    'status' => false,
-                    'message' => 'validation error',
-                    'errors' => $validateProject->errors()
-                ], 401);
-            }
         try {
             $data = Project::findOrFail($id);
-            $data->name = $request->name;
-            $data->save();
+            if(strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER' && auth()->user()->id == $data->user_id) {
+                $validateProject = Validator::make($request->all(),
+                [
+                    'name' => 'required',
+                ]);
 
-            http_response_code(200);
-            return response([
-                'message' => 'Update Successful',
-            ]);
+                if($validateProject->fails()){
+                    return response()->json([
+                        'status' => false,
+                        'message' => 'validation error',
+                        'errors' => $validateProject->errors()
+                    ], 401);
+                }
+
+                    $data->name = $request->name;
+                    $data->save();
+
+                    http_response_code(200);
+                    return response([
+                        'message' => 'Update Successful',
+                    ]);
+            }else{
+                return response([
+                    'message' => 'Only Owner Can Update Project',
+                ]);
+            }
 
         } catch (RequestException $r) {
 
             http_response_code(400);
             return response([
-                'message' => 'Data failed to be updated.',
+                'message' => 'Data failed to be updated Project.',
                 'errorCode' => 4101,
             ], 400);
         }
@@ -162,12 +201,16 @@ class ProjectController extends Controller
     {
         try {
             $data = Project::find($id);
-            $data->delete();
-
-            http_response_code(200);
-            return response([
-                'message' => 'Data successfully deleted.',
-            ]);
+            if(strtoupper(auth()->user()->userrole->role_name) == 'PRODUCT_OWNER' && auth()->user()->id == $data->user_id) {
+                 $data->delete();
+                return response([
+                    'message' => 'Data successfully deleted.',
+                ]);
+            }else{
+                return response([
+                    'message' => 'Only Owner Can Delete Project',
+                ]);
+            }
 
         } catch (RequestException $r) {
